add integration tests skeleton
- refactor epoch integration tests
