+++ integration/tests/epochs/recover_epoch/recover_epoch_efm_test.go
@@ -1,0 +1,32 @@
+package recover_epoch
+
+import (
+	"context"
+	"fmt"
+	"github.com/onflow/flow-go/model/flow"
+	"github.com/stretchr/testify/suite"
+	"testing"
+	"time"
+)
+
+func TestRecoverEpoch(t *testing.T) {
+	suite.Run(t, new(RecoverEpochSuite))
+}
+
+type RecoverEpochSuite struct {
+	Suite
+}
+
+// TestRecoverEpoch ensures that the recover_epoch transaction flow works as expected. This test will simulate the network going
+// into EFM by taking a consensus node offline before completing the DKG. While in EFM mode the test will execute the efm-recover-tx-args
+// CLI command to generate transaction arguments to submit a recover_epoch transaction, after submitting the transaction the test will
+// ensure the network is healthy.
+func (s *RecoverEpochSuite) TestRecoverEpoch() {
+	s.AwaitEpochPhase(context.Background(), 0, flow.EpochPhaseSetup, 20*time.Second, time.Second)
+	fmt.Println("in epoch phase setup")
+
+	sns := s.GetContainersByRole(flow.RoleConsensus)
+	sns[0].Pause()
+
+	// @TODO: trigger EFM manually
+}
